perf: shrink AnnotatedHit_ by reordering fields and using float for fractions
Reorder struct members (doubles, floats, int, uint16_t) to minimize
padding and use float instead of double for prefix_fraction,
suffix_fraction, and mean_error (summary statistics where 7 significant
digits is sufficient). Reduces AnnotatedHit_ from 56 to 40 bytes
(excluding AASequence/StringView), down from 72 bytes before this PR.

+++ src/openms/include/OpenMS/ANALYSIS/ID/SimpleSearchEngineAlgorithm.h
@@ -50,10 +50,11 @@ class OPENMS_DLLAPI SimpleSearchEngineAlgorithm :
     {
       StringView sequence;
       SignedSize peptide_mod_index; ///< enumeration index of the non-RNA peptide modification
+      // Layout: doubles first, then floats, then int, then uint16_t — minimizes padding
       double score = 0; ///< main score
-      double prefix_fraction = 0; ///< fraction of annotated b-ions
-      double suffix_fraction = 0; ///< fraction of annotated y-ions
-      double mean_error = 0.0; ///< mean absolute fragment mass error
+      float prefix_fraction = 0; ///< fraction of annotated b-ions
+      float suffix_fraction = 0; ///< fraction of annotated y-ions
+      float mean_error = 0.0f; ///< mean absolute fragment mass error
       int isotope_error = 0;
       uint16_t matched_b_ions = 0; ///< number of matched b-ions
       uint16_t matched_y_ions = 0; ///< number of matched y-ions
